requirements.txt
alembic==0.8.8
click==6.6
factory-boy==2.9.2
fake-factory==0.7.2
Faker==0.8.4
Flask==0.10.1
Flask-Migrate==2.0.0
Flask-Runner==2.1.1
Flask-Script==0.6.7
Flask-SQLAlchemy==2.1
Flask-Testing==0.6.1
inflection==0.3.1
ipaddress==1.0.17
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.4
MarkupSafe==0.23
marshmallow==2.10.3
multidict==2.1.2
nose==1.3.7
nose2==0.6.5
numpy==1.11.2
pandas==0.19.1
pep8==1.7.0
psycopg2==2.7.3.1
python-dateutil==2.6.1
python-editor==1.0.1
pytz==2016.10
six==1.11.0
SQLAlchemy==1.1.1
webargs==1.4.0
Werkzeug==0.11.11
